Move Schedules OnNow into schedules/lazyload namespace
Stop conflating two concepts:
* Partials are parts of a page that get included in others with a Twig
  {% include %} tag, resulting in a single HTML document.
* Lazyloads are parts of a page that get transcluded into a page at load
  time via an AJAX request. The chunk of HTML that gets trancluded has
  its own URL, thus there are two HTML documents.

The Schedules On Now block is a lazyload.
