+++ config/routes.yaml
@@ -89,7 +89,7 @@ schedules_by_network_url_key:
 
 schedules_on_now:
     path: /schedules/network/{networkUrlKey}/on-now
-    defaults: { _controller: App\Controller\Partial\SchedulesOnNowController }
+    defaults: { _controller: App\Controller\Schedules\Lazyload\OnNowController }
     requirements: { networkUrlKey: '[a-zA-Z0-9]{2,35}' }
 
 # Simple routes
